refactor(types): separate types from components
#28
